Remove unused code from `axolotl/utils/config.py`, adding validation.py back after sloppy merge
